release : 0.2.0

+++ family-moments/src/main/java/com/spring/familymoments/domain/family/FamilyController.java
@@ -5,6 +5,7 @@ import com.spring.familymoments.config.BaseResponse;
 import com.spring.familymoments.config.secret.jwt.JwtService;
 import com.spring.familymoments.domain.awsS3.AwsS3Service;
 import com.spring.familymoments.domain.family.model.*;
+import com.spring.familymoments.domain.user.AuthService;
 import com.spring.familymoments.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,6 +26,7 @@ public class FamilyController {
 
     private final FamilyService familyService;
     private final JwtService jwtService;
+    private final AuthService authService;
     @Autowired
     private final AwsS3Service awsS3Service;
 
@@ -38,7 +40,13 @@ public class FamilyController {
     @PostMapping("/family")
     public BaseResponse<PostFamilyRes> createFamily(@AuthenticationPrincipal User user,
                                                     @RequestParam(name = "representImg") MultipartFile representImg,
-                                                    @RequestPart PostFamilyReq postFamilyReq) {
+                                                    @RequestPart PostFamilyReq postFamilyReq,
+                                                    @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) {
+
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try{
 //        int owner = jwtService.getUserIdx();
             // 대표 이미지 넣기
@@ -59,8 +67,17 @@ public class FamilyController {
      */
     @ResponseBody
     @GetMapping("/{familyId}")
-    public BaseResponse<FamilyDto> getFamily(@PathVariable Long familyId) throws BaseException{
+    public BaseResponse<FamilyDto> getFamily(@PathVariable Long familyId,
+                                             @AuthenticationPrincipal User user,
+                                             @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
         //return new BaseResponse<>(familyService.getFamily(familyId));
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+        if(user == null) {
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
+        }
+
         try {
             FamilyDto familyDto = familyService.getFamily(familyId);
             return new BaseResponse<>(familyDto);
@@ -77,7 +94,12 @@ public class FamilyController {
     @ResponseBody
     @GetMapping("/{familyId}/created")
     public BaseResponse<GetFamilyCreatedNicknameRes> getFamilyCreatedNickname(@AuthenticationPrincipal User user,
-                                                                              @PathVariable Long familyId){
+                                                                              @PathVariable Long familyId,
+                                                                              @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try {
             GetFamilyCreatedNicknameRes getFamilyCreatedNicknameRes = familyService.getFamilyCreatedNickname(user, familyId);
             return new BaseResponse<>(getFamilyCreatedNicknameRes);
@@ -92,7 +114,12 @@ public class FamilyController {
      * @return BaseResponse<FamilyDto>
      */
     @GetMapping("/{familyId}/users")
-    public BaseResponse<List<GetFamilyAllRes>> getFamilyAll(@PathVariable Long familyId){
+    public BaseResponse<List<GetFamilyAllRes>> getFamilyAll(@PathVariable Long familyId,
+                                                            @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try {
             List<GetFamilyAllRes> getFamilyAllRes = familyService.getFamilyAll(familyId);
             return new BaseResponse<>(getFamilyAllRes);
@@ -107,7 +134,16 @@ public class FamilyController {
      * @return BaseResponse<FamilyDto>
      */
     @PostMapping("/inviteCode")
-    public BaseResponse<FamilyDto> getFamilyByInviteCode(@RequestBody String inviteCode){
+    public BaseResponse<FamilyDto> getFamilyByInviteCode(@RequestBody String inviteCode,
+                                                         @AuthenticationPrincipal User user,
+                                                         @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+        if(user == null) {
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
+        }
+
         try {
             FamilyDto familyDto = familyService.getFamilyByInviteCode(inviteCode);
             return new BaseResponse<>(familyDto);
@@ -124,7 +160,15 @@ public class FamilyController {
     @PostMapping("/{familyId}")
     public BaseResponse<String> inviteUser(@PathVariable Long familyId,
                                            @RequestParam List<String> userIds,
-                                           @AuthenticationPrincipal User user) throws BaseException{
+                                           @AuthenticationPrincipal User user,
+                                           @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+        if(user == null) {
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
+        }
+
         try {
             familyService.inviteUser(user, userIds, familyId);
             return new BaseResponse<>("초대 요청이 완료되었습니다.");
@@ -142,7 +186,16 @@ public class FamilyController {
      */
     @PatchMapping("/{familyId}/invite-accept")
     public BaseResponse<String> acceptFamily(@PathVariable Long familyId,
-                                             @AuthenticationPrincipal User user) throws BaseException{
+                                             @AuthenticationPrincipal User user,
+                                             @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+        if(user == null) {
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
+        }
+
         try {
             familyService.acceptFamily(user, familyId);
             return new BaseResponse<>("초대가 수락되었습니다.");
@@ -159,7 +212,12 @@ public class FamilyController {
     @PatchMapping("/{familyId}")
     public BaseResponse<String> updateUploadCycle(@AuthenticationPrincipal User user,
                                                   @PathVariable Long familyId,
-                                                  @RequestParam("uploadCycle") int uploadCycle){
+                                                  @RequestParam("uploadCycle") int uploadCycle,
+                                                  @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try {
             familyService.updateUploadCycle(user, familyId, uploadCycle);
             return new BaseResponse<>("업로드 주기가 수정되었습니다.");
@@ -174,7 +232,13 @@ public class FamilyController {
      * @return BaseResponse<String>
      */
     @DeleteMapping("/{familyId}")
-    public BaseResponse<String> deleteFamily(@AuthenticationPrincipal User user, @PathVariable Long familyId) {
+    public BaseResponse<String> deleteFamily(@AuthenticationPrincipal User user,
+                                             @PathVariable Long familyId,
+                                             @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) {
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+
         try {
             familyService.deleteFamily(user, familyId);
             return new BaseResponse<>("가족이 삭제되었습니다.");
@@ -190,7 +254,16 @@ public class FamilyController {
     @PatchMapping("/{familyId}/update")
     public BaseResponse<FamilyDto> updateFamily(@PathVariable Long familyId,
                                                 @AuthenticationPrincipal User user,
-                                                @RequestBody FamilyUpdateDto familyUpdateDto){
+                                                @RequestBody FamilyUpdateDto familyUpdateDto,
+                                                @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+        if(user == null) {
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
+        }
+
         try {
             FamilyDto resFamilyDto = familyService.updateFamily(user, familyId, familyUpdateDto);
             return new BaseResponse<>(resFamilyDto);
@@ -207,7 +280,15 @@ public class FamilyController {
      */
     @DeleteMapping("/{familyId}/withdraw")
     public BaseResponse<String> withdrawFamily(@PathVariable Long familyId,
-                                               @AuthenticationPrincipal User user){
+                                               @AuthenticationPrincipal User user,
+                                               @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+        if(user == null) {
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
+        }
+
         try {
             familyService.withdrawFamily(user, familyId);
             return new BaseResponse<>("가족에서 탈퇴되었습니다.");
@@ -225,7 +306,15 @@ public class FamilyController {
     @DeleteMapping("/{familyId}/emission")
     public BaseResponse<String> emissionFamily(@PathVariable Long familyId,
                                                @AuthenticationPrincipal User user,
-                                               @RequestParam List<String> userIds){
+                                               @RequestParam List<String> userIds,
+                                               @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
+        if(user == null) {
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
+        }
+
         try {
             familyService.emissionFamily(user, familyId, userIds);
             return new BaseResponse<>("가족에서 탈퇴되었습니다.");
