+++ family-moments/src/main/java/com/spring/familymoments/domain/user/UserController.java
@@ -2,8 +2,9 @@ package com.spring.familymoments.domain.user;
 
 import com.spring.familymoments.config.BaseException;
 import com.spring.familymoments.config.BaseResponse;
-import com.spring.familymoments.config.advice.exception.InternalServerErrorException;
+import com.spring.familymoments.config.secret.jwt.JwtService;
 import com.spring.familymoments.domain.awsS3.AwsS3Service;
+import com.spring.familymoments.domain.redis.RedisService;
 import com.spring.familymoments.domain.user.entity.User;
 import com.spring.familymoments.domain.user.model.*;
 import lombok.RequiredArgsConstructor;
@@ -18,6 +19,7 @@ import javax.mail.MessagingException;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletResponse;
 
+import java.util.Date;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -32,6 +34,8 @@ public class UserController {
     private final EmailService emailService;
     private final AwsS3Service awsS3Service;
     private final AuthService authService;
+    private final RedisService redisService;
+    private final JwtService jwtService;
 
     /**
      * 회원 가입 API
@@ -45,7 +49,7 @@ public class UserController {
     public BaseResponse<PostUserRes> createUser(@RequestPart("newUser") PostUserReq.joinUser postUserReq,
                                                 @RequestPart("profileImg") MultipartFile profileImage) throws BaseException {
         //아이디
-        if(postUserReq.getId() == null) {
+        if(postUserReq.getId() == null || postUserReq.getId().isEmpty()) {
             return new BaseResponse<>(USERS_EMPTY_USER_ID);
         }
         if(!isRegexId(postUserReq.getId())) {
@@ -61,11 +65,11 @@ public class UserController {
             return new BaseResponse<>(POST_USERS_INVALID_PW);
         }
         //이름
-        if(postUserReq.getName() == null) {
+        if(postUserReq.getName() == null || postUserReq.getName().isEmpty()) {
             return new BaseResponse<>(POST_USERS_EMPTY_NAME);
         }
         //이메일
-        if(postUserReq.getEmail() == null) {
+        if(postUserReq.getEmail() == null || postUserReq.getEmail().isEmpty()) {
             return new BaseResponse<>(POST_USERS_EMPTY_EMAIL);
         }
         if(!isRegexEmail(postUserReq.getEmail())) {
@@ -77,11 +81,14 @@ public class UserController {
             return new BaseResponse<>(POST_USERS_EXISTS_EMAIL);
         }
         //생년월일
+        if(postUserReq.getStrBirthDate() == null || postUserReq.getStrBirthDate().isEmpty()) {
+            return new BaseResponse<>(POST_USERS_EMPTY_BIRTH);
+        }
         if(!isRegexBirth(postUserReq.getStrBirthDate())) {
             return new BaseResponse<>(POST_USERS_INVALID_BIRTH);
         }
         //닉네임
-        if(postUserReq.getNickname() == null) {
+        if(postUserReq.getNickname() == null || postUserReq.getNickname().isEmpty()) {
             return new BaseResponse<>(POST_USERS_EMPTY_NICKNAME);
         }
         if(!isRegexNickName(postUserReq.getNickname())) {
@@ -106,7 +113,7 @@ public class UserController {
      * [GET] /users/check-id
      * @return BaseResponse<String>
      */
-    @GetMapping("/users/check-id")
+    @PostMapping("/users/check-id")
     public BaseResponse<String> checkDuplicateId(@RequestBody GetDuplicateUserIdReq getDuplicateUserIdReq) throws BaseException {
         try{
             if(!userService.checkDuplicateId(getDuplicateUserIdReq.getId())) {
@@ -124,7 +131,7 @@ public class UserController {
      * [GET] /users/check-email
      * @return BaseResponse<String>
      */
-    @GetMapping("/users/check-email")
+    @PostMapping("/users/check-email")
     public BaseResponse<String> checkDuplicateEmail(@RequestBody GetDuplicateUserEmailReq getDuplicateUserEmailReq) throws BaseException {
         try{
             if(!userService.checkDuplicateEmail(getDuplicateUserEmailReq.getEmail())) {
@@ -143,14 +150,14 @@ public class UserController {
      */
     @PostMapping("/users/auth/find-id")
     public BaseResponse<GetUserIdRes> findUserId(@RequestBody PostEmailReq.sendVerificationEmail sendEmailReq)
-            throws InternalServerErrorException, MessagingException, BaseException {
+            throws MessagingException, BaseException {
 
         //이름
-        if(sendEmailReq.getName() == null) {
+        if(sendEmailReq.getName() == null || sendEmailReq.getName().isEmpty()) {
             return new BaseResponse<>(POST_USERS_EMPTY_NAME);
         }
         //이메일
-        if(sendEmailReq.getEmail() == null) {
+        if(sendEmailReq.getEmail() == null || sendEmailReq.getEmail().isEmpty()) {
             return new BaseResponse<>(POST_USERS_EMPTY_EMAIL);
         }
         if(!isRegexEmail(sendEmailReq.getEmail())) {
@@ -177,7 +184,7 @@ public class UserController {
      */
     @PostMapping("/users/auth/check-id")
     public BaseResponse<String> findUserIdBeforeUpdatePwd(@RequestBody GetUserIdReq getUserIdReq)
-            throws InternalServerErrorException, MessagingException, BaseException {
+            throws MessagingException, BaseException {
         try {
             if (userService.checkDuplicateId(getUserIdReq.getUserId())) {
                 // GetUserIdRes getUserIdRes = new GetUserIdRes(id);
@@ -197,14 +204,14 @@ public class UserController {
      */
     @PostMapping("/users/auth/find-pwd")
     public BaseResponse<String> findUserPwd(@RequestBody PostEmailReq.sendVerificationEmail sendEmailReq)
-            throws InternalServerErrorException, MessagingException, BaseException {
+            throws MessagingException, BaseException {
 
         //이름
-        if(sendEmailReq.getName() == null) {
+        if(sendEmailReq.getName() == null || sendEmailReq.getName().isEmpty()) {
             return new BaseResponse<>(POST_USERS_EMPTY_NAME);
         }
         //이메일
-        if(sendEmailReq.getEmail() == null) {
+        if(sendEmailReq.getEmail() == null || sendEmailReq.getEmail().isEmpty()) {
             return new BaseResponse<>(POST_USERS_EMPTY_EMAIL);
         }
         if(!isRegexEmail(sendEmailReq.getEmail())) {
@@ -216,9 +223,9 @@ public class UserController {
                 GetUserIdRes getUserIdRes = emailService.findUserId(sendEmailReq);
                 return new BaseResponse<String>("이메일이 인증되었습니다. 새로운 비밀번호를 입력해주세요.");
             } else if(emailService.checkVerificationCode(sendEmailReq) && !emailService.checkNameAndEmail(sendEmailReq)) {
-                return new BaseResponse<>(FIND_FAIL_USER_NAME_EMAIL);
+                return new BaseResponse<>(false, FIND_FAIL_USER_NAME_EMAIL.getMessage(), HttpStatus.NOT_FOUND.value());
             } else {
-                return new BaseResponse<>(NOT_EQUAL_VERIFICATION_CODE);
+                return new BaseResponse<>(false, NOT_EQUAL_VERIFICATION_CODE.getMessage(), HttpStatus.BAD_REQUEST.value());
             }
         } catch (NoSuchElementException e) {
             return new BaseResponse<>(false, e.getMessage(), HttpStatus.NOT_FOUND.value());
@@ -231,9 +238,13 @@ public class UserController {
      * @return BaseResponse<GetProfileRes>
      */
     @RequestMapping("/users/profile")
-    public BaseResponse<GetProfileRes> readProfile(@RequestParam(value = "familyId", required = false) Long familyId, @AuthenticationPrincipal User user) {
+    public BaseResponse<GetProfileRes> readProfile(@RequestParam(value = "familyId", required = false) Long familyId,
+                                                   @AuthenticationPrincipal User user, @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) {
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
         if(user == null) {
-            return new BaseResponse<>(INVALID_USER_JWT);
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
         }
         try {
             GetProfileRes getProfileRes = userService.readProfile(user, familyId);
@@ -251,9 +262,13 @@ public class UserController {
      * @return BaseResponse<List<GetSearchUserRes>>
      */
     @GetMapping("/users")
-    public BaseResponse<List<GetSearchUserRes>> searchUser(@RequestParam(value = "keyword", required = false) String keyword, @RequestParam(value = "familyId", required = false) Long familyId, @AuthenticationPrincipal User user) {
+    public BaseResponse<List<GetSearchUserRes>> searchUser(@RequestParam(value = "keyword", required = false) String keyword, @RequestParam(value = "familyId", required = false) Long familyId,
+                                                           @AuthenticationPrincipal User user, @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) {
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
         if(user == null) {
-            return new BaseResponse<>(INVALID_USER_JWT);
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
         }
         List<GetSearchUserRes> getSearchUserRes = userService.searchUserById(keyword, familyId, user);
         return new BaseResponse<>(getSearchUserRes);
@@ -264,7 +279,15 @@ public class UserController {
      * @return BaseResponse<List<GetInvitationRes>>
      */
     @GetMapping("/users/invitation")
-    public BaseResponse<List<GetInvitationRes>> getInvitationList(@AuthenticationPrincipal User user){
+    public BaseResponse<List<GetInvitationRes>> getInvitationList(@AuthenticationPrincipal User user,
+                                                                  @RequestHeader("X-AUTH-TOKEN") String requestAccessToken){
+        if (authService.validate(requestAccessToken)) {
+            return new BaseResponse<>(INVALID_JWT);
+        }
+        if(user == null) {
+            return new BaseResponse<>(INVALID_USER_JWT);
+        }
+
         try {
             List<GetInvitationRes> getInvitationRes = userService.getInvitationList(user);
 
@@ -285,12 +308,33 @@ public class UserController {
      * @return BaseResponse<PatchProfileReqRes>
      */
     @PatchMapping("/users")
-    public BaseResponse<PatchProfileReqRes> updateProfile(@RequestParam(name = "profileImg") MultipartFile profileImg, @RequestPart PatchProfileReqRes patchProfileReqRes, @AuthenticationPrincipal User user) throws BaseException {
+    public BaseResponse<PatchProfileReqRes> updateProfile(@RequestParam(name = "profileImg") MultipartFile profileImg,
+                                                          @RequestPart PatchProfileReqRes patchProfileReqRes,
+                                                          @AuthenticationPrincipal User user, @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) throws BaseException {
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
         if(user == null) {
-            return new BaseResponse<>(INVALID_USER_JWT);
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
         }
         String fileUrl = awsS3Service.uploadImage(profileImg);
         patchProfileReqRes.setProfileImg(fileUrl);
+
+        if(patchProfileReqRes.getName() == null || patchProfileReqRes.getName().isEmpty()) { //이름 비어있으면
+            return new BaseResponse<>(POST_USERS_EMPTY_NAME);
+        }
+        if(patchProfileReqRes.getBirthdate() == null || patchProfileReqRes.getBirthdate().isEmpty()) { //생년월일 비어있으면
+            return new BaseResponse<>(POST_USERS_EMPTY_BIRTH);
+        }
+        if(!isRegexBirth(patchProfileReqRes.getBirthdate())) { //생년월일 형식 다르면
+            return new BaseResponse<>(POST_USERS_INVALID_BIRTH);
+        }
+        if(patchProfileReqRes.getNickname() == null || patchProfileReqRes.getNickname().isEmpty()) { //닉네임 비어있으면
+            return new BaseResponse<>(POST_USERS_EMPTY_NICKNAME);
+        }
+        if(!isRegexNickName(patchProfileReqRes.getNickname())) { //닉네임 형식 다르면
+            return new BaseResponse<>(POST_USERS_INVALID_NICKNAME);
+        }
 
         PatchProfileReqRes updatedUser = userService.updateProfile(patchProfileReqRes, user);
         return new BaseResponse<>(updatedUser);
@@ -301,9 +345,13 @@ public class UserController {
      * @return BaseResponse<String>
      */
     @GetMapping("/users/auth/compare-pwd")
-    public BaseResponse<String> authenticate(@RequestBody GetPwdReq getPwdReq, @AuthenticationPrincipal User user) {
+    public BaseResponse<String> authenticate(@RequestBody GetPwdReq getPwdReq,
+                                             @AuthenticationPrincipal User user, @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) {
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
         if(user == null) {
-            return new BaseResponse<>(INVALID_USER_JWT);
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
         }
         try {
             if(userService.authenticate(getPwdReq, user)) {
@@ -324,12 +372,16 @@ public class UserController {
      */
     @Transactional
     @PatchMapping("/users/modify-pwd")
-    public BaseResponse<String> updatePassword(@RequestHeader("X-AUTH-TOKEN") String requestAccessToken, @RequestBody PatchPwdReq patchPwdReq, @AuthenticationPrincipal User user, HttpServletResponse response) {
+    public BaseResponse<String> updatePassword(@RequestBody PatchPwdReq patchPwdReq,
+                                               @AuthenticationPrincipal User user, @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) {
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
         if(user == null) {
-            return new BaseResponse<>(INVALID_USER_JWT);
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
         }
         //1. 비밀번호 변경
-        if(!authenticate(new GetPwdReq(patchPwdReq.getPassword()), user).getIsSuccess()) { //비밀번호 인증
+        if(!authenticate(new GetPwdReq(patchPwdReq.getPassword()), user, requestAccessToken).getIsSuccess()) { //비밀번호 인증
             return new BaseResponse<>(FAILED_AUTHENTICATION);
         }
         if(patchPwdReq.getPassword().equals(patchPwdReq.getNewPassword())) { //newPassword와 password 일치시
@@ -360,23 +412,35 @@ public class UserController {
     @Transactional
     @PatchMapping("/users/auth/modify-pwd")
     public BaseResponse<String> updatePasswordWithoutLogin(@RequestBody PatchPwdWithoutLoginReq patchPwdWithoutLoginReq,
-                                                           @RequestParam String id) {
+                                                           @RequestParam String id) throws BaseException{
 
-        String memberEmail = emailService.getUserId(id);
+        String memberId = emailService.getUserId(id);
 
-        if(!isRegexPw(patchPwdWithoutLoginReq.getPasswordA())) {
-            return new BaseResponse<>(POST_USERS_INVALID_PW);
+        try {
+            if(userService.checkDuplicateId(memberId)) {
+                if(!isRegexPw(patchPwdWithoutLoginReq.getPasswordA()) || !isRegexPw(patchPwdWithoutLoginReq.getPasswordB())) {
+                    return new BaseResponse<>(POST_USERS_INVALID_PW);
+                }
+                if(patchPwdWithoutLoginReq.getPasswordA() == "" || patchPwdWithoutLoginReq.getPasswordA().isEmpty()) {
+                    return new BaseResponse<>(EMPTY_PASSWORD);
+                }
+                if(patchPwdWithoutLoginReq.getPasswordB() == "" || patchPwdWithoutLoginReq.getPasswordB().isEmpty()) {
+                    return new BaseResponse<>(EMPTY_PASSWORD);
+                }
+                if(!patchPwdWithoutLoginReq.getPasswordA().equals(patchPwdWithoutLoginReq.getPasswordB())) {
+                    return new BaseResponse<>(NOT_EQUAL_NEW_PASSWORD);
+                }
+
+                userService.updatePasswordWithoutLogin(patchPwdWithoutLoginReq, memberId);
+
+                return new BaseResponse<>("비밀번호가 변경되었습니다. 다시 로그인을 진행해주세요.");
+            } else {
+                return new BaseResponse<>(false, FIND_FAIL_USER_ID.getMessage(), HttpStatus.NOT_FOUND.value());
+            }
+        } catch (NoSuchElementException e) {
+            return new BaseResponse<>(false, e.getMessage(), HttpStatus.NOT_FOUND.value());
         }
-        if(patchPwdWithoutLoginReq.getPasswordB() == "") {
-            return new BaseResponse<>(EMPTY_PASSWORD);
-        }
-        if(!patchPwdWithoutLoginReq.getPasswordA().equals(patchPwdWithoutLoginReq.getPasswordB())) {
-            return new BaseResponse<>(NOT_EQUAL_NEW_PASSWORD);
-        }
 
-        userService.updatePasswordWithoutLogin(patchPwdWithoutLoginReq, memberEmail);
-
-        return new BaseResponse<>("비밀번호가 변경되었습니다. 다시 로그인을 진행해주세요.");
     }
 
     /**
@@ -397,15 +461,27 @@ public class UserController {
      */
     @Transactional
     @DeleteMapping("/users")
-    public BaseResponse<String> deleteUser(@AuthenticationPrincipal User user) {
+    public BaseResponse<String> deleteUser(@AuthenticationPrincipal User user, @RequestHeader("X-AUTH-TOKEN") String requestAccessToken) {
+        if (authService.validate(requestAccessToken)) { //유효한 사용자라 true가 반환됩니다 !!
+            return new BaseResponse<>(INVALID_JWT); //401 error : 유효한 사용자이지만, 토큰의 유효 기간이 만료됨.
+        }
         if(user == null) {
-            return new BaseResponse<>(INVALID_USER_JWT);
+            return new BaseResponse<>(INVALID_USER_JWT); //403 error : 유효한 사용자가 아님.
         }
         try {
             userService.deleteUser(user);
+            //Redis에 저장되어 있는 RT 삭제
+            String refreshTokenInRedis = redisService.getValues("RT(" + "SERVER" + "):" + user);
+            if(refreshTokenInRedis != null) {
+                redisService.deleteValues("RT(" + "SERVER" + "):" + user);
+            }
+            //Redis에 탈퇴 처리한 AT 저장
+            long expiration = jwtService.getTokenExpirationTime(requestAccessToken) - new Date().getTime();
+            redisService.setValuesWithTimeout(requestAccessToken, "delete", expiration);
+
             return new BaseResponse<>("계정을 삭제했습니다.");
         } catch (IllegalAccessException e) {
-            return new BaseResponse(false, e.getMessage(), 500);
+            return new BaseResponse<>(false, e.getMessage(), 500);
         }
     }
 }
